feat: Add DTOs for project and solution templates, refactor elevate.js to utilize new DTOs
- Created project.js for project template data transformation with methods to assign sequence numbers, format recommended roles, and format project template DTO.
- Created solution.js for solution template data transformation with a method to format solution templates from resource and program details.
- Refactored elevate.js to remove redundant template formatting functions and integrate new DTO methods for project and solution templates.
- Introduced targeting.js helper for processing targeting criteria, enhancing modularity and maintainability.
