+++ src/consumption/elevate.js
@@ -15,8 +15,6 @@ const resourceService = require('@services/resource')
 const rolloutService = require('@services/rollouts')
 const targetingHelpers = require('@helpers/targetingCriteria')
 const { Op } = require('sequelize')
-const requests = require('@generics/requests')
-const responseCode = require('@generics/http-status')
 const rolloutQueries = require('@database/queries/rollouts')
 const { ObjectId } = require('mongodb')
 const fs = require('fs')
@@ -31,6 +29,10 @@ let mongoConnection = null
 let scopeKeys = {}
 let socketInUse = false
 const resourceQueries = require('@database/queries/resources')
+const projectDTO = require('@consumption/dtos/elevate/project')
+const programDTO = require('@consumption/dtos/elevate/program')
+const solutionDTO = require('@consumption/dtos/elevate/solution')
+const targetingHelper = require('@consumption/helpers/elevate/targeting')
 
 // Define the mongoDb collection names used
 const COLLECTIONS_MAP = new Map(
@@ -61,44 +63,6 @@ const connectMongo = async (url) => {
 		return mongoConnection.getDb()
 	} catch (error) {
 		throw new Error('Error in mongo connection.')
-	}
-}
-
-/**
- * Format Project Template
- * @name formatTemplate
- * @param {Object} templateData - Project template data
- * @returns {Object} - Response contains formatted template
- */
-const formatTemplate = (templateData) => {
-	try {
-		let template = {
-			title: templateData.title,
-			tenantId: templateData.tenant_code,
-			orgId: templateData.organization_code,
-			description: templateData.objective || '',
-			keywords: utils.formatKeywords(templateData.keywords),
-			isDeleted: false,
-			createdBy: templateData.user_id,
-			updatedBy: templateData.user_id,
-			learningResources: utils.convertResources(templateData.learning_resources || []),
-			isReusable: true,
-			deleted: false,
-			status: common.PUBLISHED_STATUS,
-			externalId: utils.generateExternalId(templateData.title),
-			entityType: templateData?.entityType || '',
-			metaInformation: utils.formatProjectMetaInformation(templateData),
-			recommendedFor: [], //Initially empty
-			categories: [], //Initially empty
-			tasks: [], // Initially empty
-			taskSequence: [], // Initially empty
-			createdAt: new Date(),
-			updatedAt: new Date(),
-		}
-		return { success: true, template }
-	} catch (error) {
-		console.error('Error in formatTemplate:', error.message)
-		return { success: false, error: error.message }
 	}
 }
 
@@ -194,51 +158,6 @@ async function processCategories(categories, orgCode, tenantCode) {
 		console.error('Error in processCategories:', error.message)
 		return { success: false, error: `Failed to process categories: ${error.message}` }
 	}
-}
-
-/**
- * Converts the recommended roles for projects based on the consumption service type.
- * @name convertRecommendedRolesForProjects
- * @param {Array} recommendedFor - An array of objects containing label and value for recommended roles.
- * @returns {Object} The result object containing success status and recommended roles.
- */
-async function convertRecommendedRolesForProjects(recommendedFor) {
-	try {
-		const recommendedRoles = recommendedFor?.length
-			? recommendedFor.filter((item) => item?.label).map((item) => item.label)
-			: []
-
-		return { success: true, recommendedRoles }
-	} catch (error) {
-		return { success: false, error: `Failed to process recommeded for: ${error.message}` }
-	}
-}
-
-/**
- * Assign sequence number for task
- * @name assignSequenceNumbers
- * @returns {Object} - Response contains task object
- */
-const assignSequenceNumbers = (tasks) => {
-	/* Temporory fix start, because elevate-project doent have the observation capability in tasks now */
-	// Filter out 'observation' type tasks
-	const filteredTasks = tasks.filter((task) => task.type !== common.OBSERVATION)
-	// Sort tasks based on their current sequence number (ascending order)
-	filteredTasks.sort((a, b) => a.sequence_no - b.sequence_no)
-	let sequenceCounter = 1
-	return filteredTasks.map((task) => {
-		task.sequence_no = sequenceCounter++ // Reassign sequence number
-		return task
-	})
-	/* Temporory fix end */
-
-	// let sequenceCounter = 1
-	// return tasks.map((task) => {
-	// 	if (!task.sequence_no) {
-	// 		task.sequence_no = sequenceCounter++
-	// 	}
-	// 	return task
-	// })
 }
 
 /**
@@ -389,314 +308,6 @@ async function createTasks(tasks, templateId, templateExternalId, parentId = nul
 }
 
 /**
- * Fetch external entities from respective service
- * @name fetchExternalEntities
- * @param {Object} apiData - task data
- * @param {Array} dataToFetch - List of entities to fetch
- * @param {String} entityType - Type of the entity
- * @param {String} tenantCode - tenant code of the entities
- * @returns {Array} - Array of entity names
- */
-const fetchExternalEntities = async (apiData, dataToFetch = [], entityType, tenantCode) => {
-	let result = []
-	try {
-		if (apiData && Object.keys(apiData) && dataToFetch.length > 0) {
-			const hostEnvKey = apiData.service.replace(/-/g, '_').toUpperCase()
-			const host = process.env?.[`${hostEnvKey}_SERVICE_HOST`]
-			const serviceName = process.env?.[`${hostEnvKey}_SERVICE_NAME`]
-			const baseUrl = utils.buildUrl(host, serviceName)
-			const endPoint = utils.buildUrl(baseUrl, apiData.endPointService)
-			const bodyData = {
-				query: {
-					_id: {
-						$in: dataToFetch,
-					},
-					entityType: entityType,
-					tenantId: tenantCode,
-				},
-				projection: ['_id', 'metaInformation.name', 'entityType'],
-				mongoIdKeys: ['_id'],
-			}
-
-			const response = await requests.post(endPoint, bodyData, '', true, 'internal-access-token')
-			if (response.status == responseCode.ok) {
-				result = response?.result || []
-			}
-		}
-
-		return result
-	} catch (error) {
-		console.log(error)
-		return result
-	}
-}
-
-/**
- * Process targeting criteria
- * @name processTargetingCriteria
- * @param {Object} targetingData - Program template data
- * @returns {Object} - Response contains scope and metaInformation
- */
-const processTargetingCriteria = async (targetingData, organizationCode, tenantCode) => {
-	try {
-		let scope = {}
-		let keysToRemoveFromScope = []
-
-		// Configuration for mapping keys to data paths
-		const scopeKeyToDataPath = {
-			roles: 'professional_role',
-			sub_roles: 'professional_subroles',
-		}
-		targetingData = targetingData.map((criteria) => {
-			for (let criteriaKey of Object.keys(criteria)) {
-				let isKeyModified = false
-				// find data path if the key is modified
-				const dataPath = scopeKeyToDataPath?.[criteriaKey] || null
-				// if key is modified (i.e dataPath is not null ) and the scope is expecting multi select
-				if (dataPath && scopeKeys?.[dataPath]?.multi_select) {
-					// if key is modified and the actual data path has values
-					if (criteria?.[dataPath]?.length > 0 && criteria?.[criteriaKey]?.length > 0) {
-						criteria[dataPath] = [...criteria[dataPath], ...criteria[criteriaKey]]
-						isKeyModified = true
-						// if key is modified and the actual data path has no values or the key is not present in targeting
-					} else if (!criteria?.[dataPath] && criteria?.[criteriaKey].length > 0) {
-						criteria[dataPath] = [...criteria[criteriaKey]]
-						isKeyModified = true
-					}
-				} else if (dataPath && !scopeKeys?.[dataPath]?.multi_select) {
-					criteria[dataPath] = criteria[criteriaKey]
-					isKeyModified = true
-				}
-				if (dataPath && isKeyModified) keysToRemoveFromScope.push(criteriaKey)
-				if (
-					scopeKeys &&
-					!Object.keys(scopeKeys).includes(criteriaKey) &&
-					!keysToRemoveFromScope.includes(criteriaKey)
-				)
-					keysToRemoveFromScope.push(criteriaKey)
-			}
-			return criteria
-		})
-		// add organization into the scope by default
-		scope[`${common.SCOPE_ELEMENT_ORGANIZATIONS}`] = [organizationCode]
-		let mandatoryKeys = []
-		// iterate through the scope keys and create empty array for each key
-		// also track the mandatory keys
-		for (let scopeElement of Object.keys(scopeKeys)) {
-			scope[scopeElement] = []
-			if (scopeKeys[scopeElement]?.mandatory) {
-				mandatoryKeys.push(scopeElement)
-			}
-		}
-
-		let metaInformation = {}
-		const metaInformationKeys = [...new Set(process.env.PROGRAM_META_INFO_KEYS.split(',').map((key) => key))]
-
-		if (targetingData && Object.keys(targetingData).length > 0 && scope && Object.keys(scope).length > 0) {
-			// Iterate through each targeting criterion
-			for (let i = 0; i < targetingData.length; i++) {
-				const targeting = targetingData[i]
-				let skipTargeting = false // flag to skip further processing if 'ALL' is found
-				for (let eachTargeting of Object.keys(targeting)) {
-					const target = targeting?.[eachTargeting] || null
-					if (!Object.keys(scope).includes(eachTargeting)) scope[eachTargeting] = []
-					// check if the current scope already has 'ALL' keyword and set the flag
-					if (
-						scope[eachTargeting] == common.TARGETING_ALL ||
-						scope[eachTargeting].includes(common.TARGETING_ALL)
-					) {
-						skipTargeting = true
-					}
-					// if the particular targeting has 'ALL' keyword, ignore the processing
-					if (!skipTargeting) {
-						if (target && typeof target == common.STRING) {
-							// if the target is string , possibly we are expecting the _id of the entity.
-							// Hence push it directly making sure the value is unique
-							if (!scope[eachTargeting].includes(target)) {
-								if (scope[eachTargeting] == common.TARGETING_ALL) {
-									scope[eachTargeting] = [common.TARGETING_ALL] // if targeting is all , set the array as ["ALL"]
-								} else {
-									scope[eachTargeting].push(target)
-								}
-							}
-						} else if (target && Array.isArray(target) && target.length > 0) {
-							// if any of the element is ALL , record only ALL
-							if (target.includes(common.TARGETING_ALL)) {
-								scope[eachTargeting] = [common.TARGETING_ALL] // if targeting is all , set the array as ["ALL"]
-							} else {
-								// if the target is an array , iterate through each element
-								target.forEach((targetEntity) => {
-									// if the element inside array is string , possibly we are expecting the _id of the entity.
-									// Hence push it directly making sure the value is unique
-									if (typeof targetEntity == common.STRING)
-										if (!scope[eachTargeting].includes(targetEntity))
-											scope[eachTargeting].push(targetEntity)
-									if (typeof targetEntity == common.OBJECT) {
-										// if the element inside array is an object.
-										// check for _id or id within the object
-										const id = targetEntity?._id || targetEntity?.id || null
-										if (id && !scope[eachTargeting].includes(id)) scope[eachTargeting].push(id)
-									}
-								})
-							}
-						} else if (target && typeof target == common.OBJECT && Object.keys(target).length > 0) {
-							// if the target is an object.
-							// check for _id or id within the object.
-							const id = target?._id || target?.id || null
-							if (id && !scope[eachTargeting].includes(target)) scope[eachTargeting].push(id)
-						}
-					}
-				}
-			}
-
-			function createMetaInfo(targetingCriteria, metaInformationKeys, keyToDataPath) {
-				const metaInfo = {}
-				metaInformationKeys.forEach((key) => {
-					metaInfo[key] = new Set()
-				})
-
-				targetingCriteria.forEach((criteria) => {
-					metaInformationKeys.forEach((key) => {
-						const dataPath = keyToDataPath[key]
-						if (dataPath) {
-							const items = criteria[dataPath] || []
-							items.forEach((item) => {
-								if (item.name) {
-									metaInfo[key].add(item.name)
-								}
-							})
-						}
-					})
-				})
-
-				metaInformationKeys.forEach((key) => {
-					metaInfo[key] = Array.from(metaInfo[key])
-				})
-
-				return metaInfo
-			}
-
-			// Configuration for mapping keys to data paths
-			const keyToDataPath = {
-				state: 'state',
-				recommendedFor: 'roles',
-			}
-
-			metaInformation = createMetaInfo(targetingData, metaInformationKeys, keyToDataPath)
-		}
-		if (mandatoryKeys.length > 0) {
-			for (const key of mandatoryKeys) {
-				if (!scope[key] || scope[key].length == 0) {
-					scope[key] = [common.TARGETING_ALL]
-				}
-			}
-		}
-		if (scope) {
-			scope = Object.fromEntries(
-				Object.entries(scope).filter(([key, value]) =>
-					Array.isArray(value) ? value.length > 0 : value !== null && value !== undefined
-				)
-			)
-		}
-		if (keysToRemoveFromScope.length > 0) {
-			for (const key of keysToRemoveFromScope) {
-				scope[key] && delete scope[key]
-			}
-		}
-		return { scope, metaInformation, success: true }
-	} catch (error) {
-		console.log('Error in creating targeting : ', error)
-		return {
-			success: false,
-			error,
-		}
-	}
-}
-
-/**
- * Format Program Template
- * @name formatProgramTemplate
- * @param {Object} programData - Program template data
- * @returns {Object} - Response contains formatted template
- */
-const formatProgramTemplate = async (programData) => {
-	try {
-		let programDocument = {}
-		if (programData?.targeting_criteria) {
-			const targeting = await processTargetingCriteria(
-				programData?.targeting_criteria,
-				programData.organization_code,
-				programData.tenant_code
-			)
-			if (!targeting?.success) {
-				return {
-					success: false,
-					error: targeting?.error,
-				}
-			}
-			programDocument.scope = targeting?.scope ? targeting?.scope : {}
-			programDocument.metaInformation = targeting?.metaInformation ? targeting?.metaInformation : {}
-		}
-		programDocument.updatedAt = new Date()
-		programDocument.endDate = new Date(programData?.end_date)
-		programDocument.startDate = new Date(programData?.start_date)
-		// if the program is already published , update _id from the published_id
-		if (programData?.published_id) {
-			programDocument._id = ObjectId(programData.published_id)
-		} else {
-			let language = programData?.language
-				? programData?.resource.flatMap((resource) => {
-						return resource.languages.map((language) => {
-							return language.label
-						})
-				  })
-				: []
-
-			language = [...new Set(language)]
-			let keywords = programData?.keywords
-				? programData?.keywords
-				: programData?.resource
-				? utils.formatKeywords(programData?.resource?.keywords)
-				: []
-			keywords = [...new Set(keywords)]
-
-			programDocument = {
-				...programDocument,
-				...{
-					resourceType: [common.ROLLOUT_TYPE_PROGRAM],
-					language,
-					keywords,
-					concepts: programData?.concepts ? programData?.concepts : [],
-					components: [],
-					isAPrivateProgram: false,
-					isDeleted: false,
-					requestForPIIConsent: programData?.requestForPIIConsent ? true : false,
-					rootOrganisations: [
-						programData?.rootOrganisations ? programData?.rootOrganisations : programData?.organization?.id,
-					],
-					createdFor: [programData?.createdFor ? programData?.createdFor : programData?.organization?.id],
-					deleted: false,
-					status: common.STATUS_ACTIVE.toLowerCase(),
-					owner: programData?.created_by,
-					createdBy: programData?.created_by,
-					updatedBy: programData?.created_by,
-					externalId: utils.generateExternalId(programData?.title),
-					name: programData?.title.trim(),
-					description: programData?.resource?.objective || '',
-					createdAt: new Date(),
-					scp_reference_id: programData.resource_id,
-					orgId: programData.organization_code,
-					tenantId: programData.tenant_code,
-				},
-			}
-		}
-		return { success: true, programDocument }
-	} catch (error) {
-		console.error('Error in formatTemplate:', error.message)
-		return { success: false, error: error.message }
-	}
-}
-/**
  * Publish the project template
  * @name publishProjectTemplates
  * @param {Object} templateData - Project template data
@@ -727,13 +338,13 @@ const publishProjectTemplates = function (templateData) {
 				throw new Error('FAILED_TO_FETCH_PROJECT')
 			}
 
-			// Format the template
-			let formattedTemplate = formatTemplate({ ...projectData })
-			if (!formattedTemplate.success || !formattedTemplate?.template) {
+			// Format the template using DTO
+			const formattedTemplate = projectDTO.formatProjectTemplateDTO({ ...projectData })
+			if (!formattedTemplate.success || !formattedTemplate?.data) {
 				throw new Error('FAILED_TO_FORMAT_TEMPLATE')
 			}
 
-			let template = formattedTemplate.template
+			let template = formattedTemplate.data
 
 			projectsMongoConnection = projectsMongoConnection
 				? projectsMongoConnection
@@ -770,18 +381,16 @@ const publishProjectTemplates = function (templateData) {
 
 			//process recommededFor
 			if (projectData.recommended_for?.length > 0) {
-				let recommededForResponse = await convertRecommendedRolesForProjects(projectData.recommended_for)
-				if (!recommededForResponse.success) {
+				let recommededForResponse = projectDTO.formatRecommendedRoles(projectData.recommended_for)
+				if (!recommededForResponse.success || !recommededForResponse?.data) {
 					throw new Error('FAILED_TO_FETCH_RECOMMENDED_FOR')
 				}
-				template.recommendedFor = recommededForResponse?.recommendedRoles
+				template.recommendedFor = recommededForResponse?.data
 			}
 
 			// Insert the template into the database
 			const templateCollection = projectsMongoConnection.collection(COLLECTIONS_MAP.get('TEMPLATES'))
-			const result = await templateCollection.insertOne(template)
-
-			// Validate the result of the template creation
+			const result = await templateCollection.insertOne(template) // Validate the result of the template creation
 			if (!result || !result.insertedId) {
 				throw new Error('FAILED_TO_CREATE_TEMPLATE')
 			}
@@ -789,7 +398,7 @@ const publishProjectTemplates = function (templateData) {
 			const templateId = result.insertedId
 
 			// Process and Create Tasks
-			const processedTasks = assignSequenceNumbers(projectData.tasks || [])
+			const processedTasks = projectDTO.assignSequenceNumbers(projectData.tasks || [])
 			const taskCreationResponse = await createTasks(
 				processedTasks,
 				templateId,
@@ -919,10 +528,11 @@ async function updateProgram(programId, updateTemplate) {
  * @return {Object} updateBody - resource update body
  */
 async function updateSolutionTemplate(resource) {
-	const targeting = await processTargetingCriteria(
+	const targeting = await targetingHelper.processTargetingCriteria(
 		resource?.targeting_criteria,
 		resource.organization_code,
-		resource.tenant_code
+		resource.tenant_code,
+		scopeKeys
 	)
 	if (!targeting?.success) {
 		return {
@@ -1712,60 +1322,14 @@ const createSolutions = async (resourceDetails, programDetails, userToken) => {
 		// solution to rollout if map
 		let solutionRolloutMap = {}
 
-		const endDate = new Date(programDetails?.end_date)
-		const startDate = new Date(programDetails?.start_date)
 		resourceDetails.forEach((resource) => {
-			// create solutions template
-			const solutionTemplate = {
-				resourceType: [common.SOLUTIONS_RESOURCE_TYPE[resource.type]],
-				language: resource?.languages ? resource?.languages.map((language) => language.label) : [],
-				keywords: resource?.keywords ? utils.formatKeywords(resource?.keywords) : [],
-				concepts: resource?.concepts ? resource?.concepts : [],
-				themes: resource?.themes ? resource?.themes : [],
-				flattenedThemes: resource?.flattenedThemes ? resource?.flattenedThemes : [],
-				entities: resource?.entities ? resource?.entities : [],
-				registry: resource?.registry ? resource?.registry : [],
-				isRubricDriven: resource?.isRubricDriven ? true : false,
-				scp_reference_id: resource?.resource_id,
-				enableQuestionReadOut: resource?.enableQuestionReadOut ? true : false,
-				captureGpsLocationAtQuestionLevel: resource?.captureGpsLocationAtQuestionLevel ? true : false,
-				isAPrivateProgram: false,
-				allowMultipleAssessemts: resource?.allowMultipleAssessemts ? true : false,
-				isDeleted: false,
-				pageHeading: 'Domains',
-				minNoOfSubmissionsRequired: resource?.minNoOfSubmissionsRequired
-					? resource?.minNoOfSubmissionsRequired
-					: 1,
-				rootOrganisations: resource?.organization
-					? resource?.organization.map((organization) => organization.id)
-					: [],
-				createdFor: resource?.organization ? resource?.organization.map((organization) => organization.id) : [],
-				deleted: false,
-				name: resource?.title,
-				programExternalId: programDetails.externalId,
-				entityType: resource?.entityType ? resource?.entityType : null,
-				type: common.SOLUTIONS_TYPE[resource.type] ? common.SOLUTIONS_TYPE[resource.type] : null,
-				subType: common.SOLUTIONS_TYPE[resource.type] ? common.SOLUTIONS_TYPE[resource.type] : null,
-				isReusable: false,
-				externalId: utils.generateUniqueId(),
-				programId: programDetails._id,
-				programName: programDetails.name,
-				programDescription: programDetails.description,
-				description: resource?.description ? resource.description : programDetails.description,
-				status: common.STATUS_ACTIVE.toLowerCase(),
-				updatedAt: new Date(),
-				createdAt: new Date(),
-				scope: programDetails.scope,
-				projectTemplateId: resource._id,
-				updatedBy: programDetails.created_by,
-				author: programDetails.created_by,
-				endDate,
-				startDate,
-				creator: programDetails.created_by,
-				orgId: programDetails.orgId,
-				tenantId: programDetails.tenantId,
-				referenceFrom: programDetails.referenceFrom ? programDetails.referenceFrom : '',
+			// create solutions template using DTO
+			const formattedSolution = solutionDTO.formatSolutionTemplate(resource, programDetails)
+			if (!formattedSolution.success || !formattedSolution?.data) {
+				throw new Error(`Failed to format solution template: ${formattedSolution?.error || 'Unknown error'}`)
 			}
+
+			const solutionTemplate = formattedSolution.data
 
 			solutionRolloutMap[solutionTemplate.externalId] = resource.rolloutId
 
@@ -1908,51 +1472,6 @@ const createSolutions = async (resourceDetails, programDetails, userToken) => {
 	}
 }
 
-const orderSolutionsInProgram = (resourceWithInProgram) => {
-	let solutionOrderList = resourceWithInProgram.map((item) => {
-		let res = {
-			id: item.id,
-		}
-		if (item?.published_id) res._id = ObjectId(item.published_id)
-		if (item?.order) res.order = item.order
-		return res
-	})
-
-	const usedOrders = new Set()
-
-	// First, process items with explicit orders
-	for (let i = 0; i < resourceWithInProgram.length; i++) {
-		const item = resourceWithInProgram[i]
-		if (item.order != null) {
-			let ord = item.order
-			while (usedOrders.has(ord)) {
-				ord++
-			}
-			solutionOrderList[i].order = ord
-			usedOrders.add(ord)
-		}
-	}
-
-	// Then, process items without explicit orders (null or undefined)
-	for (let i = 0; i < resourceWithInProgram.length; i++) {
-		const item = resourceWithInProgram[i]
-		if (item.order == null) {
-			let ord = i + 1
-			while (usedOrders.has(ord)) {
-				ord++
-			}
-			solutionOrderList[i].order = ord
-			usedOrders.add(ord)
-		}
-	}
-
-	return solutionOrderList.reduce((acc, item) => {
-		acc[item.id] = { order: item.order }
-		if (item._id) acc[item.id]._id = item._id
-		return acc
-	}, {})
-}
-
 /**
  * Publish the Program
  * @name publishProjectTemplates
@@ -1979,14 +1498,14 @@ const publishProgram = function async(programData) {
 			const isProgramResource = rolloutDetails.resource_type === common.RESOURCE_TYPE_PROGRAM
 			const programResourceTableId = isProgramResource ? rolloutDetails.resource_id : null
 
-			// Format the program template
-			let formattedTemplate = await formatProgramTemplate(rolloutDetails)
+			// Format the program template using DTO
+			const formattedTemplate = await programDTO.formatProgramTemplateDTO(rolloutDetails, scopeKeys)
 
-			if (!formattedTemplate.success) {
+			if (!formattedTemplate.success || !formattedTemplate?.data) {
 				throw new Error('FAILED_TO_FORMAT_TEMPLATE')
 			}
 
-			let template = formattedTemplate.programDocument
+			let template = formattedTemplate.data
 
 			let programResourceRolloutMap = {}
 			let programResourceIds = []
@@ -2051,7 +1570,7 @@ const publishProgram = function async(programData) {
 				throw new Error('NO_RESOURCE_ADDED')
 			}
 			let solutionIds = []
-			let solutionOrderMap = orderSolutionsInProgram(resourceWithInProgram)
+			let solutionOrderMap = programDTO.orderSolutionsInProgram(resourceWithInProgram)
 			let resourceToUpdate = []
 
 			for (const resource of resourceWithInProgram) {
@@ -2127,10 +1646,11 @@ const publishProgram = function async(programData) {
 							)
 						}
 
-						const targeting = await processTargetingCriteria(
+						const targeting = await targetingHelper.processTargetingCriteria(
 							fetchDetails?.result?.targeting_criteria,
 							fetchDetails?.result?.organization_code,
-							fetchDetails?.result?.tenant_code
+							fetchDetails?.result?.tenant_code,
+							scopeKeys
 						)
 						if (!targeting?.success) {
 							throw new Error(
